+++ app/microstrain_mag_cal_apply/cli.hpp
@@ -1,0 +1,21 @@
+#pragma once
+
+#include <CLI/CLI.hpp>
+
+
+namespace cli
+{
+    struct ProgramArgs
+    {
+        explicit ProgramArgs(char** argv);
+
+        CLI::App app{};
+
+        std::filesystem::path calibration_filepath;
+        std::string port_name;
+        std::uint32_t baudrate = 115200;
+        bool overwrite = false;
+    };
+
+    void displayFailedConnectionInformation(std::string_view port_name, std::uint32_t baudrate);
+}